refactor(google-meet): split CLI modules
